CSharpDB 4.5.0: complete SQL type semantics

+++ tests/CSharpDB.Migration.Files.Tests/JsonMigrationCatalogTests.cs
@@ -166,7 +166,7 @@ public sealed class JsonMigrationCatalogTests
             AssertMapping(
                 plan,
                 4,
-                DbType.Integer,
+                DbType.Decimal,
                 MigrationMappingClassification.LosslessReencoded);
             AssertMapping(
                 plan,
